View screen improvements:
- Buttons to filtering logs
- Activity info reduced
- Card buttons with lower case and primary color
- All drawable background unified
- DpToPx and PxToDp fixed

+++ library/src/main/java/es/rafaco/inappdevtools/library/view/icons/IconDrawable.java
@@ -108,7 +108,7 @@ public class IconDrawable extends Drawable {
      * @return The current IconDrawable for chaining.
      */
     public IconDrawable sizeDp(int size) {
-        return sizePx(UiUtils.dpToPx(context, size));
+        return sizePx((int) UiUtils.dpToPx(context, size));
     }
 
     /**
